requirements.txt
jupyterlab
python-dotenv
langchain==0.0.316
openai==0.28.1